refactor article plugin into components
